pyproject.toml
[tool.poetry]
name = "wikipedia-mobile"
version = "0.1.0"
description = ""
authors = ["Maksim Kudaev <kudaev.m@nakukop.com>"]
readme = "README.md"
packages = [{include = "wikipedia_mobile"}]

[tool.poetry.dependencies]
python = "^3.11"
Appium-Python-Client = "2.6.1"
selene = "^2.0.0b17"
pytest = "^7.1.3"
allure-pytest = "^2.10.0"
python-dotenv = "^0.21.0"
pydantic = "^1.10.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = " -vv -s --alluredir=allure-results "
